+++ src/main/resources/routes
@@ -29,14 +29,15 @@ GET         /publish/company/:id/apply-for-code               controllers.Report
 
 GET         /publish/company/:id/file/page/:formName          controllers.LongFormController.show(formName: LongFormName, id: CompaniesHouseId)
 POST        /publish/company/:id/file/page/:formName          controllers.LongFormController.post(formName: LongFormName, id: CompaniesHouseId)
-GET         /publish/company/:id/review/long                  controllers.LongFormController.showReview(id:CompaniesHouseId)
 
 GET         /publish/company/:id/file/short                   controllers.ShortFormController.show(id:CompaniesHouseId)
 POST        /publish/company/:id/file/short                   controllers.ShortFormController.post(id:CompaniesHouseId)
 
-POST        /publish/company/:id/review/long                  controllers.LongFormController.postReview(id:CompaniesHouseId)
-GET         /publish/company/:id/review/short                 controllers.ShortFormController.showReview(id:CompaniesHouseId)
-POST        /publish/company/:id/review/short                 controllers.ShortFormController.postReview(id:CompaniesHouseId)
+GET         /publish/company/:id/review/long                  controllers.LongFormReviewController.showReview(id:CompaniesHouseId)
+POST        /publish/company/:id/review/long                  controllers.LongFormReviewController.postReview(id:CompaniesHouseId)
+
+GET         /publish/company/:id/review/short                 controllers.ShortFormReviewController.showReview(id:CompaniesHouseId)
+POST        /publish/company/:id/review/short                 controllers.ShortFormReviewController.postReview(id:CompaniesHouseId)
 
 GET         /publish/:reportId/success                        controllers.ConfirmationController.showConfirmation(reportId:ReportId)
 GET         /report/:reportId                                 controllers.ReportController.view(reportId: ReportId)
